added ResponseContract

+++ Applicationn/Behaviour/Institution/DeleteInstitutionCommand.cs
@@ -1,9 +1,12 @@
-﻿using FluentValidation;
+﻿using Application.Utils;
+using Domain.Constants;
+using Domain.Entities;
+using FluentValidation;
 using MediatR;
 
 namespace Application.Behaviour.Institution
 {
-    public record DeleteInstitutionCommand : IRequest
+    public record DeleteInstitutionCommand : IRequest<ResponseContract<Unit>>
     {
         public Guid Id { get; set; }
         public bool IsArchive { get; set; }
@@ -18,11 +21,19 @@ namespace Application.Behaviour.Institution
         }
     }
 
-    public sealed class DeleteInstitutionCommandHandler(IInstitutionRepository repository) : IRequestHandler<DeleteInstitutionCommand>
+    public sealed class DeleteInstitutionCommandHandler(IInstitutionRepository repository) : IRequestHandler<DeleteInstitutionCommand, ResponseContract<Unit>>
     {
-        public async Task Handle(DeleteInstitutionCommand request, CancellationToken cancellationToken)
+        public async Task<ResponseContract<Unit>> Handle(DeleteInstitutionCommand request, CancellationToken cancellationToken)
         {
+            InstitutionEntity institution = await repository.GetByIdAsync(request.Id);
+            if (institution == null)
+            {
+                return new ResponseContract<Unit>(ErrorCodes.InstitutionNotFound);
+            }
+
             await repository.DeleteAsync(request.Id);
+
+            return new ResponseContract<Unit>(Unit.Value);
         }
     }
 }
